# api/openapi-alpha.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VerdeAI v5.1 API Scaffold",
    "version": "5.1.0",
    "description": "Mock backend contract for future VerdeAI vision, plant overlay, tester-page handoff, future composition, rendering and reporting services."
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Check backend health",
        "responses": {
          "200": {
            "description": "Backend is reachable"
          }
        }
      }
    },
    "/api/analyse": {
      "post": {
        "summary": "Analyse uploaded property context",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageDataUrl": {
                    "type": "string"
                  },
                  "propertyType": {
                    "type": "string",
                    "description": "Includes v5.1 values such as needs-review and under-building."
                  },
                  "preference": {
                    "type": "string"
                  },
                  "constraint": {
                    "type": "string",
                    "description": "Main problem selected by the tester, e.g. shade-dark, access-awkward, privacy-gap, maintenance-drag, water-risk."
                  },
                  "budget": {
                    "type": "string"
                  },
                  "maintenance": {
                    "type": "string"
                  },
                  "postcode": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  },
                  "starterCue": {
                    "type": "string",
                    "description": "Optional v5.1 starter clue such as shade, access, tired, utility, or edge."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property DNA result"
          }
        }
      }
    },
    "/api/futures": {
      "post": {
        "summary": "Generate possible property futures",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "propertyType": {
                    "type": "string"
                  },
                  "preference": {
                    "type": "string"
                  },
                  "constraint": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Future concept list"
          }
        }
      }
    },
    "/api/render": {
      "post": {
        "summary": "Prepare a mock render request or future provider render call",
        "responses": {
          "200": {
            "description": "Mock render result with provider, future, prompt, and estimated cost fields"
          }
        },
        "description": "v5.1 render scaffold. Real providers are not connected by default; this endpoint documents the future provider proxy contract.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "string",
                    "enum": [
                      "none",
                      "replicate-flux",
                      "openai-image",
                      "stability-ai"
                    ]
                  },
                  "futureId": {
                    "type": "string"
                  },
                  "count": {
                    "type": "number"
                  },
                  "prompt": {
                    "type": "string"
                  },
                  "overlayLabels": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/report": {
      "post": {
        "summary": "Generate a report from project state",
        "responses": {
          "200": {
            "description": "Generated report"
          }
        }
      }
    }
  }
}